Prefix shard map IDs for _users and _replicator
The scheme of using the DbName as the shard map ID will result
in shard maps that cannot be edited for the _users and _replicator
DBs. This commit teaches mem3 to prefix shard map IDs for those
databases.

Note that because shard map IDs are also generated in fabric this
commit also requires:

    couchdb-fabric@497a5ad4a291e13e33dc0570d176c4a344d8c37e

COUCHDB-2422